requirements.txt
absl-py==2.2.2
blas==1.0
brotli-python==1.0.9
bzip2==1.0.8
ca-certificates==2025.2.25
certifi==2025.1.31
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.2
cuda-cccl==12.8.90
cuda-cccl_win-64==12.8.90
cuda-cudart==11.8.89
cuda-cudart-dev==11.8.89
cuda-cupti==11.8.87
cuda-libraries==11.8.0
cuda-libraries-dev==11.8.0
cuda-nvrtc==11.8.89
cuda-nvrtc-dev==11.8.89
cuda-nvtx==11.8.86
cuda-profiler-api==12.8.90
cuda-runtime==11.8.0
cuda-version==12.8
cycler==0.12.1
einops==0.8.1
filelock==3.17.0
fonttools==4.57.0
freeglut==3.4.0
freetype==2.13.3
fsspec==2025.3.2
ftfy==6.3.1
giflib==5.2.2
gmp==6.3.0
gmpy2==2.2.1
grpcio==1.71.0
idna==3.7
imageio==2.37.0
intel-openmp==2023.1.0
jinja2==3.1.6
jpeg==9e
kiwisolver==1.4.8
kornia==0.8.0
kornia-rs==0.1.8
lazy-loader==0.4
lcms2==2.16
lerc==4.0.0
libcublas==11.11.3.6
libcublas-dev==11.11.3.6
libcufft==10.9.0.58
libcufft-dev==10.9.0.58
libcurand==10.3.9.90
libcurand-dev==10.3.9.90
libcusolver==11.4.1.48
libcusolver-dev==11.4.1.48
libcusparse==11.7.5.86
libcusparse-dev==11.7.5.86
libdeflate==1.22
libffi==3.4.4
libjpeg-turbo==2.0.0
libnpp==11.8.0.86
libnpp-dev==11.8.0.86
libnvjpeg==11.9.0.86
libnvjpeg-dev==11.9.0.86
libpng==1.6.39
libtiff==4.7.0
libuv==1.48.0
libwebp==1.3.2
libwebp-base==1.3.2
lz4-c==1.9.4
markdown==3.8
markupsafe==3.0.2
matplotlib==3.10.1
mkl==2023.1.0
mkl-service==2.4.0
mkl_fft==1.3.11
mkl_random==1.2.8
mpc==1.3.1
mpfr==4.2.1
mpmath==1.3.0
natsort==8.4.0
networkx==3.4.2
numpy==2.0.1
numpy-base==2.0.1
openai-clip==1.0.1
opencv-python==4.11.0.86
openjpeg==2.5.2
openssl==3.0.16
packaging==25.0
pandas==2.2.3
pillow==11.1.0
pip==25.0
protobuf==6.30.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
pysocks==1.7.1
python==3.11.11
python-dateutil==2.9.0.post0
pytorch==2.5.1
pytorch-cuda==11.8
pytorch-mutex==1.0
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
sqlite==3.45.3
sympy==1.13.1
tbb==2021.8.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tifffile==2025.3.30
tk==8.6.14
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.113
ultralytics-thop==2.0.14
urllib3==2.3.0
vc==14.42
vs2015_runtime==14.42.34433
wcwidth==0.2.13
werkzeug==3.1.3
wheel==0.45.1
win_inet_pton==1.1.0
xz==5.6.4
yaml==0.2.5
zlib==1.2.13
zstd==1.5.6